refactor: extract upgradeCommandExecutor from preUpgradeCommandsExecutor
