+++ lib/workers/repository/update/branch/execute-pre-upgrade-commands.ts
@@ -1,16 +1,14 @@
 import is from '@sindresorhus/is';
-import { mergeChildConfig } from '../../../../config';
 import { GlobalConfig } from '../../../../config/global';
 import { addMeta, logger } from '../../../../logger';
 import type { ArtifactError } from '../../../../modules/manager/types';
-import { exec } from '../../../../util/exec';
 import { localPathIsFile, writeLocalFile } from '../../../../util/fs';
 import type { FileChange } from '../../../../util/git/types';
-import { regEx } from '../../../../util/regex';
-import { sanitize } from '../../../../util/sanitize';
-import { compile } from '../../../../util/template';
 import type { BranchConfig, BranchUpgradeConfig } from '../../../types';
-import { updateUpdatedArtifacts } from './execute-upgrade-commands';
+import {
+  updateUpdatedArtifacts,
+  upgradeCommandExecutor,
+} from './execute-upgrade-commands';
 
 export interface PreUpgradeCommandsExecutionResult {
   updatedArtifacts: FileChange[];
@@ -55,44 +53,14 @@ export async function preUpgradeCommandsExecutor(
       }
 
       for (const cmd of commands) {
-        if (
-          allowedUpgradeCommands!.some((pattern) => regEx(pattern).test(cmd))
-        ) {
-          try {
-            const compiledCmd = allowUpgradeCommandTemplating
-              ? compile(cmd, mergeChildConfig(config, upgrade))
-              : cmd;
-
-            logger.trace({ cmd: compiledCmd }, 'Executing pre-upgrade task');
-            const execResult = await exec(compiledCmd, {
-              cwd: GlobalConfig.get('localDir'),
-            });
-
-            logger.debug(
-              { cmd: compiledCmd, ...execResult },
-              'Executed pre-upgrade task'
-            );
-          } catch (error) {
-            artifactErrors.push({
-              lockFile: upgrade.packageFile,
-              stderr: sanitize(error.message),
-            });
-          }
-        } else {
-          logger.warn(
-            {
-              cmd,
-              allowedUpgradeCommands,
-            },
-            'Pre-upgrade task did not match any on allowedUpgradeCommands list'
-          );
-          artifactErrors.push({
-            lockFile: upgrade.packageFile,
-            stderr: sanitize(
-              `Pre-upgrade command '${cmd}' has not been added to the allowed list in allowedUpgradeCommands`
-            ),
-          });
-        }
+        const commandError = await upgradeCommandExecutor(
+          allowedUpgradeCommands ?? [],
+          cmd,
+          allowUpgradeCommandTemplating,
+          config,
+          upgrade
+        );
+        artifactErrors.concat(commandError);
       }
 
       updatedArtifacts = await updateUpdatedArtifacts(
